requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
colorama==0.4.5
Django==4.0.6
django-crispy-forms==1.14.0
django-filter==22.1
djangorestframework==3.13.1
idna==3.3
Jinja2==3.1.2
Markdown==3.4.1
MarkupSafe==2.1.1
Pillow==9.2.0
pytz==2022.1
PyYAML==6.0
requests==2.28.1
sqlparse==0.4.2
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.10
